+++ src/tiffIO.cpp
@@ -159,8 +159,8 @@ tiffIO::tiffIO(char *fname, DATA_TYPE newtype) {
 		nodata = new float;
 		*((float*) nodata) = (float) GDALGetRasterNoDataValue(bandh, NULL);
 	} else if (datatype == LONG_TYPE) {
-		nodata = new long;
-		*((long*) nodata) = (long) GDALGetRasterNoDataValue(bandh, NULL);
+		nodata = new int32_t;
+		*((int32_t*) nodata) = (int32_t) GDALGetRasterNoDataValue(bandh, NULL);
 	}
 
 }
@@ -189,8 +189,8 @@ tiffIO::tiffIO(char *fname, DATA_TYPE newtype, void* nd, const tiffIO &copy) {
 		nodata = new float;
 		*((float*) nodata) = *((float*) nd);
 	} else if (datatype == LONG_TYPE) {
-		nodata = new long;
-		*((long*) nodata) = *((long*) nd);
+		nodata = new int32_t;
+		*((int32_t*) nodata) = *((int32_t*) nd);
 	}
 
 	totalX = copy.totalX;
@@ -287,7 +287,7 @@ void tiffIO::write(long xstart, long ystart, long numRows, long numCols, void* s
 			
 			else if (datatype == LONG_TYPE)
 			
-				GDALSetRasterNoDataValue(bandh, (double) *((long*) nodata));
+				GDALSetRasterNoDataValue(bandh, (double) *((int32_t*) nodata));
 
 			//int n2;
